Added support for DPI to screen Get DPI correctly on Windows7 and Windows 8.1 (screen still needs to support multiple monitors but the DPI system supports it) Support for rotated sprites from texture packer TTF font exporter working now (including distance field support) Debug console print in reslience now uses TTF arial font Various warning fixes distance field 2D GPU program and shaders Font loader now binary in place (no extra processing) with multiple region support for larger non continous fonts

+++ src/apps/gfx_test/voxtree.inl
@@ -9,7 +9,7 @@ Tree<Derived>::Tree( const Core::AABB& _box ) :
 	boundingBox( _box ),
 	nodeTiles( 1 )
 {
-	uint32_t rootIndex = nodeTiles.alloc();
+	uintptr_t rootIndex = nodeTiles.alloc();
 	CORE_ASSERT( rootIndex == 0 );
 	for( int i = 0;i < 8; ++i ) {
 		nodeTiles.get(0).nodes[i].type = NodeType::EMPTY;
